+++ lib/dart_ipc.dart
@@ -1,4 +1,2 @@
-library dart_ipc;
-
 export 'src/dart_ipc.dart';
 export 'src/dart_ipc_unix.dart';