+++ lib/src/dart_ipc_unix.dart
@@ -3,19 +3,19 @@ import 'dart:io';
 import 'dart_ipc_platform_interface.dart';
 
 class DartIpcUnix extends DartIpcPlatform {
-  static void registerWith() {
-    DartIpcPlatform.instance = DartIpcUnix();
-  }
-
   @override
   Future<ServerSocket> bind(String path) {
     return ServerSocket.bind(
-        InternetAddress(path, type: InternetAddressType.unix), 0);
+      InternetAddress(path, type: InternetAddressType.unix),
+      0,
+    );
   }
 
   @override
   Future<Socket> connect(String path) {
     return Socket.connect(
-        InternetAddress(path, type: InternetAddressType.unix), 0);
+      InternetAddress(path, type: InternetAddressType.unix),
+      0,
+    );
   }
 }